ref(onboarding): Split apple onboarding docs
