+++ static/app/gettingStartedDocs/apple/apple/index.tsx
@@ -1,4 +1,4 @@
-import docs from './macos';
+import docs from 'sentry/gettingStartedDocs/apple/macos';
 
 // apple is the legacy platform key for apple-macos
 // So we simply re-export the docs from apple-macos
