+++ static/app/gettingStartedDocs/apple/macos/onboarding.spec.tsx
@@ -6,7 +6,7 @@ import {textWithMarkupMatcher} from 'sentry-test/utils';
 
 import {ProductSolution} from 'sentry/components/onboarding/gettingStartedDoc/types';
 
-import docs from './macos';
+import docs from '.';
 
 describe('apple-macos onboarding docs', () => {
   it('renders docs correctly', async () => {
